[tests] Split up too large DataMapperTests.cpp into multiple smaller segments
