+++ src/tests/DataMapper/CreateTests.cpp
@@ -1,0 +1,128 @@
+
+// SPDX-License-Identifier: Apache-2.0
+
+#include "../Utils.hpp"
+
+#include <Lightweight/Lightweight.hpp>
+
+#include <reflection-cpp/reflection.hpp>
+
+#include <catch2/catch_session.hpp>
+#include <catch2/catch_template_test_macros.hpp>
+#include <catch2/catch_test_macros.hpp>
+#include <catch2/matchers/catch_matchers.hpp>
+#include <catch2/matchers/catch_matchers_floating_point.hpp>
+
+#include <string_view>
+
+using namespace std::string_view_literals;
+using namespace std::string_literals;
+using namespace Lightweight;
+
+struct RecordWithDefaults
+{
+    Field<uint64_t, PrimaryKey::ServerSideAutoIncrement> id;
+    Field<SqlAnsiString<30>> name1 { "John Doe" };
+    Field<std::optional<SqlAnsiString<30>>> name2 { "John Doe" };
+    Field<bool> boolean1 { true };
+    Field<bool> boolean2 { false };
+    Field<std::optional<int>> int1 { 42 };
+    Field<std::optional<int>> int2 {};
+};
+
+struct RecordWithNoDefaults
+{
+    Field<uint64_t, PrimaryKey::ServerSideAutoIncrement> id;
+    Field<SqlAnsiString<30>> name1;
+    Field<std::optional<SqlAnsiString<30>>> name2;
+    Field<bool> boolean1;
+    Field<bool> boolean2;
+    Field<std::optional<int>> int1;
+    Field<std::optional<int>> int2;
+
+    static constexpr std::string_view TableName = RecordTableName<RecordWithDefaults>;
+};
+
+// NOLINTBEGIN(bugprone-unchecked-optional-access)
+
+TEST_CASE_METHOD(SqlTestFixture, "Create table with default values", "[DataMapper]")
+{
+    auto dm = DataMapper();
+    dm.CreateTable<RecordWithDefaults>();
+
+    auto record = RecordWithDefaults {};
+    dm.Create(record);
+
+    auto const actual = dm.QuerySingle<RecordWithNoDefaults>(record.id).value();
+    CHECK(actual.id == record.id);
+    CHECK(actual.name1 == record.name1);
+    CHECK(actual.boolean1 == record.boolean1);
+    CHECK(actual.boolean2 == record.boolean2);
+    CHECK(actual.int1 == record.int1);
+    CHECK(actual.int2 == record.int2);
+}
+
+struct TestRecord
+{
+    Field<uint64_t, PrimaryKey::AutoAssign> id {};
+    Field<SqlAnsiString<30>> comment;
+
+    std::weak_ordering operator<=>(TestRecord const& other) const = default;
+};
+
+std::ostream& operator<<(std::ostream& os, TestRecord const& record)
+{
+    return os << DataMapper::Inspect(record);
+}
+
+TEST_CASE_METHOD(SqlTestFixture, "PrimaryKey: AutoAssign", "[DataMapper]")
+{
+    auto dm = DataMapper();
+    dm.CreateTable<TestRecord>();
+
+    auto record = TestRecord { .comment = "Hello, World!" };
+    dm.Create(record);
+    INFO("Created record: " << DataMapper::Inspect(record));
+    auto const queriedRecord = dm.QuerySingle<TestRecord>(record.id).value();
+    REQUIRE(queriedRecord == record);
+
+    auto record2 = TestRecord { .comment = "Hello, World! 2" };
+    dm.Create(record2);
+    INFO("Created record: " << DataMapper::Inspect(record2));
+    auto const queriedRecord2 = dm.QuerySingle<TestRecord>(record2.id).value();
+    REQUIRE(queriedRecord2 == record2);
+
+    REQUIRE(record.id != record2.id);
+}
+
+struct MultiPkRecord
+{
+    Field<SqlAnsiString<32>, PrimaryKey::AutoAssign> firstName;
+    Field<SqlAnsiString<32>, PrimaryKey::AutoAssign> lastName;
+
+    constexpr std::weak_ordering operator<=>(MultiPkRecord const& other) const = default;
+};
+
+TEST_CASE_METHOD(SqlTestFixture, "Table with multiple primary keys", "[DataMapper]")
+{
+    auto dm = DataMapper {};
+
+    dm.CreateTable<MultiPkRecord>();
+
+    auto record = MultiPkRecord { .firstName = "John", .lastName = "Doe" };
+    dm.Create(record);
+
+    {
+        auto const _ = ScopedSqlNullLogger {}; // Suppress the error message, as we are testing for it
+        CHECK_THROWS_AS(dm.CreateExplicit(MultiPkRecord { .firstName = "John", .lastName = "Doe" }), SqlException);
+    }
+
+    auto queriedRecords = dm.Query<MultiPkRecord>().All();
+
+    CHECK(queriedRecords.size() == 1);
+    auto const& queriedRecord = queriedRecords.at(0);
+    INFO("Queried record: " << DataMapper::Inspect(queriedRecord));
+    CHECK(queriedRecord == record);
+}
+
+// NOLINTEND(bugprone-unchecked-optional-access)
